Enhance user creation and editing forms with additional fields and validation rules
